# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Eiendomsmuligheter Platform API
  description: |-
    Offisiell API-dokumentasjon for Eiendomsmuligheter Platform 2.0
    
    Dette API-et gir tilgang til alle funksjonene i Eiendomsmuligheter Platform, inkludert:
    * Eiendomsanalyse med AlterraML
    * 3D-visualisering med TerravisionJS
    * Kommuneintegrasjon med CommuneConnect
    * Økonomi- og bærekraftsvurderinger
    
    Mer informasjon finnes på [https://eiendomsmuligheter.no](https://eiendomsmuligheter.no)
  contact:
    email: support@eiendomsmuligheter.no
  license:
    name: Proprietær
    url: https://eiendomsmuligheter.no/license
  version: 2.0.0
servers:
  - url: https://api.eiendomsmuligheter.no/api/v1
    description: Produksjonsserver
  - url: https://staging.eiendomsmuligheter.no/api/v1
    description: Testserver
  - url: http://localhost:8000/api
    description: Lokal utviklingsserver
...
paths:
...
  /visualization/terrain/generate:
    post:
      tags:
        - visualization
      summary: Generer terrengmodell
      description: Genererer en 3D-terrengmodell basert på eiendomsdata
      operationId: generateTerrain
      requestBody:
        description: Terrenggenereringsforespørsel
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerrainGenerationRequest'
        required: true
      responses:
        '200':
          description: Vellykket generering av terreng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerrainGenerationResponse'
        '400':
          description: Ugyldig forespørsel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Serverfeil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /visualization/building/generate:
    post:
      tags:
        - visualization
      summary: Generer bygningsmodell
      description: Genererer en 3D-bygningsmodell basert på spesifikasjoner
      operationId: generateBuilding
      requestBody:
        description: Bygningsgenereringsforespørsel
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildingGenerationRequest'
        required: true
      responses:
        '200':
          description: Vellykket generering av bygning
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingGenerationResponse'
        '400':
          description: Ugyldig forespørsel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Serverfeil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    TerrainGenerationRequest:
      type: object
      required:
        - property_id
        - width
        - depth
      properties:
        property_id:
          type: string
          description: Eiendoms-ID
          example: "1234567"
        width:
          type: number
          format: float
          description: Terrengbredde i meter
          example: 100.0
        depth:
          type: number
          format: float
          description: Terrengdybde i meter
          example: 100.0
        resolution:
          type: integer
          description: Oppløsning for terrengkart
          example: 128
          default: 128
        include_surroundings:
          type: boolean
          description: Inkluder omkringliggende terreng
          example: true
          default: true
        include_buildings:
          type: boolean
          description: Inkluder bygninger
          example: true
          default: true
        texture_type:
          type: string
          description: Teksturtype (satellite, map, hybrid)
          example: "satellite"
          default: "satellite"
    TerrainGenerationResponse:
      type: object
      required:
        - heightmap_url
        - texture_url
        - metadata
        - bounds
      properties:
        heightmap_url:
          type: string
          description: URL til høydekart
          example: "/api/static/heightmaps/heightmap_1234567_abcd1234.png"
        texture_url:
          type: string
          description: URL til teksturkart
          example: "/api/static/textures/texture_1234567_abcd1234.jpg"
        metadata:
          type: object
          additionalProperties: true
          description: Metadata for terrenget
        bounds:
          type: object
          properties:
            north:
              type: number
              format: float
              example: 59.95
            south:
              type: number
              format: float
              example: 59.94
            east:
              type: number
              format: float
              example: 10.76
            west:
              type: number
              format: float
              example: 10.75
          description: Geografiske grenser
    BuildingGenerationRequest:
      type: object
      required:
        - property_id
        - building_type
        - floors
        - width
        - depth
        - height
      properties:
        property_id:
          type: string
          description: Eiendoms-ID
          example: "1234567"
        building_type:
          type: string
          description: Bygningstype
          example: "residential"
        floors:
          type: integer
          description: Antall etasjer
          example: 2
        width:
          type: number
          format: float
          description: Bygningsbredde i meter
          example: 10.0
        depth:
          type: number
          format: float
          description: Bygningsdybde i meter
          example: 10.0
        height:
          type: number
          format: float
          description: Bygningshøyde i meter
          example: 6.0
        roof_type:
          type: string
          description: Taktype
          example: "flat"
          default: "flat"
        style:
          type: string
          description: Bygningsstil
          example: "modern"
          default: "modern"
        colors:
          type: object
          additionalProperties:
            type: string
          description: Fargepalett
          example: {"walls": "#FFFFFF", "roof": "#FF0000"}
    BuildingGenerationResponse:
      type: object
      required:
        - model_url
        - thumbnail_url
        - metadata
      properties:
        model_url:
          type: string
          description: URL til 3D-modell
          example: "/api/static/models/building_1234567_abcd1234.glb"
        thumbnail_url:
          type: string
          description: URL til miniatyrbilde
          example: "/api/static/textures/building_thumb_1234567_abcd1234.jpg"
        metadata:
          type: object
          additionalProperties: true
          description: Metadata for bygningen
...
    ErrorResponse:
      type: object
      required:
        - status_code
        - detail
      properties:
        status_code:
          type: integer
          description: HTTP-statuskode
          example: 400
        detail:
          type: string
          description: Feilbeskrivelse
          example: "Ugyldig forespørsel"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-autentiseringstoken 